web/requirements.txt
asgiref==3.6.0
boto3==1.26.80
botocore==1.29.80
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-dotenv==1.4.2
django-storages==1.13.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
